feat(scripts/utils): split out functions from postbuild.js
